# Improve encode performance