Hi. I have looked at issue #1 and find out that the current htmlEncode can be improved in performance. First i have tried to include html-entities, which will work but i was not brave enogh to include it, because it bundles a lot of stuff into the views bundle. Also htmlEncode is a public API of the library and i don't wanted to change it's produced encoding, because it might break user tests. So i ended up by creating a "own" implementation again, to speed things up.

### Micro-Benchmarks
The speed improvement highly depends on the input and the underlying javascript engine. I have prepared some micro-benchmarks for you, that shows the new version should be faster. Just click to run it, these benchmarks will only run the htmlEncode function with some input.

- [encode attribute, no escape](https://jsben.ch/KTc48) 
- [encode attribute, with escapes](https://jsben.ch/FY2dX)
- [encode content, no escape](https://jsben.ch/OegRF)
- [encode content, with escapes](https://jsben.ch/WTirT)

### Complate-cpp Benchmarks
I have also run complate-cpp rendering benchmarks with the current and the optimized implementation. I'm using a Intel i7-12700KF which runs an Ubuntu 20.04 on WSL2. All results are mean values of all iterations, i have taken best of five runs each, rounded to one decimal.

| Test                         | Current        | Optimized    |
| -------------------- | ------------: | ------------: |
| QuickJS, no todos    |         92.7 us |        60.0 us |
| QuickJS, 10 todos    |          1.2 ms |      771.1 us |
| QuickJS, 100 todos  |        11.5 ms |         7.5 ms |
| V8, no todos            |        10.7 us |          9.1 us |
| V8, 10 todos            |      149.4 us |      123.0 us |
| V8, 100 todos          |       1.44 ms |       1.26 ms |

To run the complate-cpp benchmarks by yourself on Linux, assuming you have the dependencies installed on your system.
```shell
mkdir build
cd build
cmake -DBUILD_TESTS=on ..
make
test/lib/complate-tests QuickJsRendererBenchmark
test/lib/complate-tests V8RendererBenchmark
```