@tmehnert 2021-12-02
I have done a new version based on the algorithm of  [html-entities](https://github.com/mdevils/html-entities). I would really like to present you a easy oneliner, but it seems to that the loop based version deliver more stable performance across various engines.

### Complate-Cpp Benchmarks
If ran the Benchmarks again with our new algorithms. Loop-Based is the one i have commited.

| Test                         | Without Fastpath | Loop-Based  |
| -------------------- | -----------------: | ------------: |
| QuickJS, no todos    |                66.4 us |        61.3 us |
| QuickJS, 10 todos    |              828.5 us |   773.8.1 us |
| QuickJS, 100 todos  |                 8.0 ms |         7.5 ms |
| V8, no todos            |                 9.6 us |          9.0 us |
| V8, 10 todos            |             138.1 us |      125.2 us |
| V8, 100 todos          |              1.37 ms |       1.21 ms |

I have also updated the [online benchmark](https://jsben.ch/xhVX4).