@FND 2021-12-07
As noted earlier, I've tweaked a few details in a5830ea2cfc7c0b3adc5e8458e47a0aaaf644105 (see commit message); if you're fine with that, I think this is ready to go.